requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
discord.py==1.5.0
idna==2.10
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pycodestyle==2.6.0
pylint==2.6.0
python-dotenv==0.14.0
requests==2.24.0
rope==0.17.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
yarl==1.6.0
